large refactoring for new POC, implemented logic to fetch header and persist with simpler db schema (#10)
