+++ commands/rpc/rpc.go
@@ -6,7 +6,7 @@ import (
 
 // RpcCmd RPC service commands
 var RpcCmd = &cobra.Command{
-	Use:   "rpc",
+	Use:   "http",
 	Short: "RPC Service commands",
 	Long:  `The RPC Service expose a RPC endpoint compatible with the CometBFT RPC endpoint.`,
 	Run: func(cmd *cobra.Command, args []string) {
